% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNRD-package.R
\docType{package}
\name{easyNRD-package}
\alias{easyNRD}
\alias{easyNRD-package}
\title{easyNRD: Tools for HCUP-NRD Workflows}
\description{
Provides lazy, SQL-translatable tools for NRD label augmentation, episode
consolidation, and readmission linkage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AliSalman-et-al/easyNRD}
  \item Report bugs at \url{https://github.com/AliSalman-et-al/easyNRD/issues}
}

}
\author{
\strong{Maintainer}: Ali Salman \email{alisalman.et.al@gmail.com}

}
\keyword{internal}
