% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNRD-package.R
\docType{package}
\name{easyNRD-package}
\alias{easyNRD}
\alias{easyNRD-package}
\title{easyNRD: Discharge-level tools for HCUP-NRD workflows}
\description{
Provides lazy, SQL-translatable tools for discharge-level NRD ingestion,
condition flagging, readmission linkage, methodology-safe multi-year pooling,
export, and survey design preparation for HCUP-NRD data years 2016 and
later.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AliSalman-et-al/easyNRD}
  \item Report bugs at \url{https://github.com/AliSalman-et-al/easyNRD/issues}
}

}
\author{
\strong{Maintainer}: Ali Salman \email{alisalman.et.al@gmail.com}

}
\keyword{internal}
