% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNRD-package.R
\docType{package}
\name{easyNRD-package}
\alias{easyNRD}
\alias{easyNRD-package}
\title{easyNRD: Tools for HCUP-NRD Workflows}
\description{
Provides lazy, SQL-translatable tools for NRD label augmentation, episode
consolidation, and readmission linkage.
}
\author{
\strong{Maintainer}: Ali Salman \email{alisalman.et.al@gmail.com}

}
\keyword{internal}
