% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{nrd_demographics}
\alias{nrd_demographics}
\title{Return standard NRD demographic variables}
\usage{
nrd_demographics()
}
\value{
A character vector of demographic variable names.
}
\description{
Return standard NRD demographic variables
}
\examples{
nrd_demographics()
}
