% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{nrd_hospital_vars}
\alias{nrd_hospital_vars}
\title{Return standard NRD hospital structural variables}
\usage{
nrd_hospital_vars()
}
\value{
A character vector of hospital variable names.
}
\description{
Return standard NRD hospital structural variables
}
\examples{
nrd_hospital_vars()
}
